pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Alan Kleber <alan.kleber@scalefast.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.2"
psycopg = {extras = ["binary"], version = "^3.1.17"}
djangorestframework = "^3.14.0"
django-ninja = "^1.1.0"
pillow = "^10.2.0"
django-widget-tweaks = "^1.5.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
